# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Jeu des valeurs — API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sessions/{sessionId}/cartes:
    get:
      operationId: listCartesSession
      tags: [cartes]
      summary: Liste les cartes sélectionnées pour une session
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Cartes de la session
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CarteSession"
    post:
      operationId: addCarteSession
      tags: [cartes]
      summary: Ajoute une carte (du catalogue ou personnalisée)
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarteSessionInput"
      responses:
        "201":
          description: Carte ajoutée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarteSession"
        "400":
          description: Données invalides
  /sessions/{sessionId}/cartes/{carteSessionId}:
    patch:
      operationId: updateCarteSession
      tags: [cartes]
      summary: Modifie une carte (libellé, description, valeurs confirmées)
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/CarteSessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarteSessionUpdate"
      responses:
        "200":
          description: Carte mise à jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarteSession"
        "404":
          description: Carte introuvable
    delete:
      operationId: removeCarteSession
      tags: [cartes]
      summary: Retire une carte de la session
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/CarteSessionId"
      responses:
        "204":
          description: Carte retirée
        "404":
          description: Carte introuvable

  # ─── Réponses (collisions) ────────────────────────────────────────────────────
...
  /sessions/{sessionId}/constellation:
    get:
      operationId: getConstellation
      tags: [constellation]
      summary: Calcule et retourne la constellation personnelle (déterministe)
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Constellation calculée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Constellation"
        "404":
          description: Session introuvable
  /sessions/{sessionId}/progres:
    get:
      operationId: getProgres
      tags: [sessions]
      summary: Retourne un résumé de la progression de la session
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Progression
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Progres"
components:
  parameters:
    SessionId:
      in: path
      name: sessionId
      required: true
      schema:
        type: string
    CarteSessionId:
      in: path
      name: carteSessionId
      required: true
      schema:
        type: number
...
  schemas:
...
    CarteSession:
      type: object
      required: [id, sessionId, famille, label, valeursConfirmées, estPersonnalisée]
      properties:
        id:
          type: number
        sessionId:
          type: string
        catalogueCarteId:
          type: ["number", "null"]
        famille:
          type: string
          enum: [lignes_rouges, horizons, tresors]
        label:
          type: string
        description:
          type: ["string", "null"]
        valeursConfirmées:
          type: array
          items:
            type: string
        valeursSuggérées:
          type: array
          items:
            type: string
        valeurPrincipale:
          type: ["string", "null"]
        estPersonnalisée:
          type: boolean
        creeLe:
          type: string
          format: date-time
    CarteSessionInput:
      type: object
      required: [famille, label]
      properties:
        catalogueCarteId:
          type: ["number", "null"]
        famille:
          type: string
          enum: [lignes_rouges, horizons, tresors]
        label:
          type: string
        description:
          type: ["string", "null"]
        valeursSuggérées:
          type: array
          items:
            type: string
        valeursConfirmées:
          type: array
          items:
            type: string
        estPersonnalisée:
          type: boolean
        valeurPrincipale:
          type: ["string", "null"]
    CarteSessionUpdate:
      type: object
      properties:
        label:
          type: string
        description:
          type: ["string", "null"]
        valeursConfirmées:
          type: array
          items:
            type: string
        valeurPrincipale:
          type: ["string", "null"]

    # ─── Réponses ──────────────────────────────────────────────────────────────
...
    TendanceValeur:
      type: object
      required: [valeur, scoreNet, importanceApparente, incertitude, stabilite, couverture, statutProtege, contextesExplores, preferencesParContexte, totalCollisions, victoiresA, victoiresB, incertitudes, abandonnes, difficulteMoyenne, certitudeMoyenne, territoireInexplore]
      properties:
        valeur:
          type: string
        scoreNet:
          type: number
          description: Transformation de l'importance apparente sur l'intervalle -1 à 1
        importanceApparente:
          type: number
          description: Moyenne des préférences Beta(1,1) pondérées, de 0 à 1
        incertitude:
          type: number
        stabilite:
          type: number
        couverture:
          type: number
        statutProtege:
          type: boolean
        contextesExplores:
          type: array
          items:
            type: string
        preferencesParContexte:
          type: array
          items:
            type: object
            required: [contexte, importanceApparente, preuvesEffectives]
            properties:
              contexte:
                type: string
              importanceApparente:
                type: number
              preuvesEffectives:
                type: number
        totalCollisions:
          type: number
        victoiresA:
          type: number
        victoiresB:
          type: number
        incertitudes:
          type: number
          description: ca_depend + je_ne_sais_pas
        abandonnes:
          type: number
          description: passer
        difficulteMoyenne:
          type: ["number", "null"]
        certitudeMoyenne:
          type: ["number", "null"]
        territoireInexplore:
          type: boolean
          description: Vrai si aucune collision résolue pour cette valeur
    TensionObservee:
      type: object
      required: [valeurA, valeurB, totalCollisions, incertitudes, difficulteMoyenne, probabiliteA, incertitude, estForte]
      properties:
        valeurA:
          type: string
        valeurB:
          type: string
        totalCollisions:
          type: number
        incertitudes:
          type: number
        difficulteMoyenne:
          type: ["number", "null"]
        probabiliteA:
          type: number
        incertitude:
          type: number
        estForte:
          type: boolean
    ObservationConstellation:
      type: object
      required: [id, texte, type, valeursConcernees, reponsesSources]
      properties:
        id:
          type: string
        texte:
          type: string
          description: Formulation prudente (ex. "Dans les situations explorées jusqu'ici…")
        type:
          type: string
          enum: [tendance, tension, territoire_inexplore, stabilite, couverture, valeur_protegee, point_bascule]
        valeursConcernees:
          type: array
          items:
            type: string
        reponsesSources:
          type: array
          items:
            type: number
          description: IDs des réponses qui appuient cette observation
    Constellation:
      type: object
      required: [sessionId, version, tendances, tensions, observations, couverture, stabilite, versionCalcul]
      properties:
        sessionId:
          type: string
        version:
          type: number
        tendances:
          type: array
          items:
            $ref: "#/components/schemas/TendanceValeur"
        tensions:
          type: array
          items:
            $ref: "#/components/schemas/TensionObservee"
        observations:
          type: array
          items:
            $ref: "#/components/schemas/ObservationConstellation"
        couverture:
          type: number
          description: Proportion de collisions possibles explorées (0 à 1)
        stabilite:
          type: number
          description: Cohérence observée lors des répétitions comparables (0 à 1)
        versionCalcul:
          type: number
          description: Version du moteur de calcul utilisé

    # ─── Progression ───────────────────────────────────────────────────────────
    Progres:
      type: object
      required: [sessionId, etapeCourante, nombreCartes, nombreValeurs, nombreCollisions, nombreReponses, prochaineDilemme]
      properties:
        sessionId:
          type: string
        etapeCourante:
          type: string
          enum: [accueil, selection_cartes, confirmation_valeurs, collisions, constellation]
        nombreCartes:
          type: number
        nombreValeurs:
          type: number
        nombreCollisions:
          type: number
          description: Nombre de collisions possibles (combinaisons C(n,2) des valeurs)
        nombreReponses:
          type: number
          description: Nombre de collisions auxquelles une réponse a été donnée
        prochaineDilemme:
          type: ["object", "null"]
          properties:
            valeurA:
              type: string
            valeurB:
              type: string
            dilemmeId:
              type: ["number", "null"]
            texte:
              type: ["string", "null"]
            contexte:
              type: ["string", "null"]
            pivotDimension:
              type: ["string", "null"]