# lib/api-zod/src/generated/types/carteSession.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Jeu des valeurs — API
 * OpenAPI spec version: 0.1.0
 */
import type { CarteSessionFamille } from './carteSessionFamille';

export interface CarteSession {
  id: number;
  sessionId: string;
  /** @nullable */
  catalogueCarteId?: number | null;
  famille: CarteSessionFamille;
  label: string;
  /** @nullable */
  description?: string | null;
  valeursConfirmées: string[];
  valeursSuggérées?: string[];
  /** @nullable */
  valeurPrincipale?: string | null;
  estPersonnalisée: boolean;
  creeLe?: Date;
}