# lib/api-zod/src/generated/types/carteSessionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Jeu des valeurs — API
 * OpenAPI spec version: 0.1.0
 */
import type { CarteSessionInputFamille } from './carteSessionInputFamille';

export interface CarteSessionInput {
  /** @nullable */
  catalogueCarteId?: number | null;
  famille: CarteSessionInputFamille;
  label: string;
  /** @nullable */
  description?: string | null;
  valeursSuggérées?: string[];
  valeursConfirmées?: string[];
  estPersonnalisée?: boolean;
  /** @nullable */
  valeurPrincipale?: string | null;
}