# lib/api-zod/src/generated/types/carteSessionUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Jeu des valeurs — API
 * OpenAPI spec version: 0.1.0
 */

export interface CarteSessionUpdate {
  label?: string;
  /** @nullable */
  description?: string | null;
  valeursConfirmées?: string[];
  /** @nullable */
  valeurPrincipale?: string | null;
}