# lib/api-zod/src/generated/types/tendanceValeur.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Jeu des valeurs — API
 * OpenAPI spec version: 0.1.0
 */

export interface TendanceValeur {
  valeur: string;
  /** Transformation de l'importance apparente sur l'intervalle -1 à 1 */
  scoreNet: number;
  /** Moyenne des préférences Beta(1,1) pondérées, de 0 à 1 */
  importanceApparente: number;
  incertitude: number;
  stabilite: number;
  couverture: number;
  statutProtege: boolean;
  contextesExplores: string[];
  preferencesParContexte: {
    contexte: string;
    importanceApparente: number;
    preuvesEffectives: number;
  }[];
  totalCollisions: number;
  victoiresA: number;
  victoiresB: number;
  /** ca_depend + je_ne_sais_pas */
  incertitudes: number;
  /** passer */
  abandonnes: number;
  /** @nullable */
  difficulteMoyenne: number | null;
  /** @nullable */
  certitudeMoyenne: number | null;
  /** Vrai si aucune collision résolue pour cette valeur */
  territoireInexplore: boolean;
}