# lib/api-zod/src/generated/types/tensionObservee.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Jeu des valeurs — API
 * OpenAPI spec version: 0.1.0
 */

export interface TensionObservee {
  valeurA: string;
  valeurB: string;
  totalCollisions: number;
  incertitudes: number;
  /** @nullable */
  difficulteMoyenne: number | null;
  probabiliteA: number;
  incertitude: number;
  estForte: boolean;
}